[KAN-91] refactor : 내가 만든 모임 페이지 컴포넌트 로직 분리
