+++ src/app/(main)/mypage/created/page.tsx
@@ -1,40 +1,10 @@
 'use client';
 
-import getGatherings from '@/app/api/actions/gatherings/getGatherings';
+import { useUserCreated } from '@/hooks/useUserCreated';
 import GatheringList from './_component/GatheringList';
-import { GatheringsListData } from '@/types/data.type';
-import { useUser } from '@/app/(auth)/context/UserContext';
-import { useEffect, useState } from 'react';
-import { set } from 'zod';
-import { DATA_LIST } from './mockData';
 
 const CreatedPage = () => {
-  const { user } = useUser();
-  const [gatheringsList, setGatheringsList] = useState<GatheringsListData[]>(
-    [],
-  );
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchGatheringsData = async () => {
-      if (user) {
-        setIsLoading(true);
-
-        try {
-          const gatherings = await getGatherings({
-            createdBy: String(user.id),
-          });
-          setGatheringsList(gatherings);
-        } catch (error) {
-          console.error('Error fetching data:', error);
-        } finally {
-          setIsLoading(false); // 요청이 끝나면 로딩 상태 해제
-        }
-      }
-    };
-
-    fetchGatheringsData();
-  }, [user]);
+  const { gatheringsList, isLoading } = useUserCreated();
 
   return (
     <>
